Fix issue with ember-lodash when using Babel polyfill on Phantom.
When lodash detects that `Object.defineProperty` has been modified
by `core-js` polyfills, it intentionally exports `undefined` as its
default export (the rest of `lodash` knows how to handle this).

However, when `loader.js` detects that `exports.default === undefined`
it automatically sets `exports.default = exports` (which is mega WAT).

This causes the rest of the detection system that `lodash` has in place
around its `_defineProperty` exporting `undefined` as default to completely
fall apart. Since it expects that module to either export a function or
`undefined`, when it sees that `_defineProperty.default` is truthy it
simply invokes (what it thinks is a function).  Due to the behavior of
`loader.js`, this means that we are essentially doing:

```js
let sadface = {};
sadface();
```

Which, low and behold, when ran on Phantom emits the following error:

```
Object is not a constructor (evaluating '(0, _lodash_defineProperty['default'])')
```

Thanks to @rondale-sc for helping me debug this.